+++ test/__init__.py
@@ -1,0 +1,5 @@
+import sys
+import os
+
+current_path = os.path.dirname(os.path.abspath(__file__))
+sys.path.insert(0, current_path + '/../')